visualize/graph.py
import svgwrite

class Graph:
    def __init__(self, directed=False):
        self.nodes = set()
        self.edges = {}
        self.directed = directed

    def add_edge(self, u, v):
        self.nodes.add(u)
        self.nodes.add(v)
        if u in self.edges:
            self.edges[u].append(v)
        else:
            self.edges[u] = [v]
        if not self.directed:
            if v in self.edges:
                self.edges[v].append(u)
            else:
                self.edges[v] = [u]
                self.edges[v] = [u]



def main():
    # Create a list of edges for a tree with 7 nodes
    edges = {
        1: [2, 3],
        2: [3, 4, 5],
        3: [6, 7]
    }

if __name__ == '__main__':
    # Create a list of edges for a tree with 7 nodes
    edges = {
        1: [2, 3],
        2: [3, 4, 5],
        3: [6, 7]
    }

    # Create an instance of the Graph class with 7 nodes and the list of edges
    graph = Graph(7, edges)

    # Generate an SVG image of the tree and save it to a file
    graph.renderSvg('graph.svg')

class Visualizer:
    def __init__(self):
        pass

    def visualize(self, graph):
        # calculate the position of each nodes
        positions = []
        for i, u in enumerate(graph.nodes):
            x = i * 30
            y = 50
            self.position.push((x, y))